+++ src/onegov.notice/HISTORY.rst
@@ -1,7 +1,10 @@
 Changelog
 ---------
 
-- Allows filter notices by a search term.
+- Allows to filter notices by a search term.
+  [msom]
+
+- Allows to filter notices by user IDs.
   [msom]
 
 0.0.1 (2017-07-14)